separate search for one and for multiple choice
