+++ design-system/uilib/src/components/dpdhl-search/index.ts
@@ -1,2 +1,2 @@
 export { default as DPDHLSearch } from './dpdhl-search.svelte'
-export { default as DPDHLSelectItem } from './dpdhl-select-item.svelte'
+export { default as DPDHLSearchItem } from './dpdhl-search-item.svelte'
