+++ design-system/uilib/src/components/index.ts
@@ -13,3 +13,4 @@ export * from './dpdhl-ui-shell'
 export * from './dpdhl-chip'
 export * from './dpdhl-typography'
 export * from './dpdhl-select'
+export * from './dpdhl-search'
